+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - programs: Added callback-enabled instructions for order.
   - Added the `create_order_v2` instruction.
   - Added the `close_order_v2` instrcution.
+- programs: Added the `gmsol-competition` program.
 
 ### Changed
 
@@ -50,7 +51,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - programs: Implemented `Default` for `Glv` and made the `store` field public.
-- programs: Added the `gmsol-competition` program.
 - programs: Introduced a separate `impl_fixed_map!` macro that implements fixed map functionality without defining the corresponding struct.
 - programs: Introduced a separated `impl_flags!` macro that implements flag map functionality without defining the container.
 - model: Re-exported `num_traits`.